remove brownie.cli.utils, move color to brownie.utils
