+++ brownie/test/output.py
@@ -3,7 +3,7 @@
 import json
 from pathlib import Path
 
-from brownie.cli.utils import color
+from brownie.utils import color
 from brownie.network.state import TxHistory
 
 COVERAGE_COLORS = [(0.8, "bright red"), (0.9, "bright yellow"), (1, "bright green")]
